requirements.txt
pybit
numpy
python-dotenv
psutil
pytest
pytest-asyncio
cryptography
pyyaml